Remove unused code. Move some code to data_utils.py DF done
