+++ scripts/data_utils.py
@@ -2,6 +2,7 @@ import glob
 import json
 import os
 
+import numpy as np
 from PIL import Image
 
 S2S_ORIGINAL_CATEGORIES = [
@@ -174,3 +175,31 @@ def create_image_info(
     }
 
     return image
+
+
+def _resize_thumbnail(im: Image, target_image_size: tuple):
+    im.thumbnail(target_image_size, Image.ANTIALIAS)
+    new_image = Image.new("RGB", target_image_size, (255, 255, 255))
+    new_image.paste(
+        im,
+        (
+            int((target_image_size[0] - im.size[0]) / 2),
+            int((target_image_size[1] - im.size[1]) / 2),
+        ),
+    )
+
+    return new_image
+
+
+def crop_single_bbox(image, bbox, target_image_size: tuple):
+    img_arr = np.array(image)
+    bbox = np.asarray(bbox)
+    bbox_int = bbox.astype(np.int32)
+    x1, y1, bbox_w, bbox_h = bbox_int[:4]
+    x2 = x1 + bbox_w
+    y2 = y1 + bbox_h
+    cut_arr = img_arr[y1:y2, x1:x2]  # Cut-out the instance detected
+    im = Image.fromarray(cut_arr)
+    cut_arr = _resize_thumbnail(im, target_image_size)
+
+    return cut_arr
